pyproject.toml
[project]
name = "paper-to-notion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "feedparser>=6.0.12",
    "flake8>=7.3.0",
    "google-genai>=1.38.0",
    "ipykernel>=6.30.1",
    "notion-client>=2.5.0",
    "openai>=1.108.1",
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "tkcalendar>=1.6.1",
    "ujson>=5.11.0",
]
